# backend/openapi.yml
openapi: 3.0.0
info:
  title: MEH
  version: 1.0.0
  description: REST API for the MEH application
servers:
  - url: /api
paths:
...
  /content:
    post:
      summary: Create new content
      description: Create a new content item for the current user
      operationId: createContent
      tags:
        - content
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ContentCreateRequest"
      responses:
        201:
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Content"
        400:
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AppError"
        401:
          $ref: "#/components/responses/Unauthorized"
        500:
          $ref: "#/components/responses/InternalServerError"
    get:
      summary: Get all content
      description: Retrieve all content items belonging to the current user
      operationId: getContentItems
      tags:
        - content
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Content"
        401:
          $ref: "#/components/responses/Unauthorized"
        500:
          $ref: "#/components/responses/InternalServerError"
  /content/{id}:
    get:
      summary: Get content by ID
      description: Retrieve a specific content item by its unique identifier
      operationId: getContentById
      tags:
        - content
      parameters:
        - name: id
          in: path
          description: Content unique identifier
          required: true
          schema:
            type: string
            format: uuid
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Content"
        401:
          $ref: "#/components/responses/Unauthorized"
        403:
          $ref: "#/components/responses/Forbidden"
        404:
          description: Not Found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AppError"
        500:
          $ref: "#/components/responses/InternalServerError"
    put:
      summary: Update content
      description: Update an existing content item
      operationId: updateContent
      tags:
        - content
      parameters:
        - name: id
          in: path
          description: Content unique identifier
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ContentUpdateRequest"
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Content"
        400:
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AppError"
        401:
          $ref: "#/components/responses/Unauthorized"
        403:
          $ref: "#/components/responses/Forbidden"
        404:
          description: Not Found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AppError"
        500:
          $ref: "#/components/responses/InternalServerError"
    delete:
      summary: Delete content
      description: Delete (soft-delete) a content item
      operationId: deleteContent
      tags:
        - content
      parameters:
        - name: id
          in: path
          description: Content unique identifier
          required: true
          schema:
            type: string
            format: uuid
      responses:
        204:
          description: No Content
        401:
          $ref: "#/components/responses/Unauthorized"
        403:
          $ref: "#/components/responses/Forbidden"
        404:
          description: Not Found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AppError"
        500:
          $ref: "#/components/responses/InternalServerError"
components:
  responses:
    InternalServerError:
      description: Internal Server Error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/AppError"
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/AppError"
    Forbidden:
      description: Forbidden
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/AppError"
  schemas:
    AppError:
      type: object
      properties:
        code:
          type: string
        api_message:
          type: string
        status_code:
          type: integer
          format: int32
        values:
          type: object
          additionalProperties: true
      required:
        - code
        - status_code
...
    Content:
      type: object
      properties:
        id:
          type: string
          format: uuid
          description: Content unique identifier
        owner_id:
          type: string
          format: uuid
          description: ID of the user who owns this content
        created_at:
          type: string
          format: date-time
          description: Creation timestamp
        updated_at:
          type: string
          format: date-time
          description: Last update timestamp
        type:
          type: string
          description: Type of content
        data:
          type: string
          description: Content data
      required: [id, owner_id, created_at, updated_at, type, data]
    ContentCreateRequest:
      type: object
      properties:
        type:
          type: string
          description: Type of content
        data:
          type: string
          description: Content data
      required: [type, data]
    ContentUpdateRequest:
      type: object
      properties:
        type:
          type: string
          description: Type of content
        data:
          type: string
          description: Content data